[Updated #98] Changed SimulatorStream to Abstract Class
To fix this issue, we need to have the client address in a stream
so that we can respond to the client correctly. The code has been
refactored into an abstract class and the client address property
has been added to the parent.

### Next Steps:
We need to set the client address when a request is received and
have the DuplicatePacketStream send a duplicate request.